refactor(6697): rename sdk direct probe fixture support
