+++ crates/kamn-e2e-harness/src/drivers/sdk_direct_tests/support.rs
@@ -1,7 +1,7 @@
 #[path = "support/env_support.rs"]
 mod env_support;
-#[path = "support/script_fixture_support.rs"]
-mod script_fixture_support;
+#[path = "support/probe_fixture_support.rs"]
+mod probe_fixture_support;
 
 pub(crate) use env_support::*;
-pub(crate) use script_fixture_support::*;
+pub(crate) use probe_fixture_support::*;
